+++ back/src/main/java/com/baba/back/baby/dto/CreateInviteCodeResponse.java
@@ -1,4 +1,4 @@
-package com.baba.back.baby.invitation.dto;
+package com.baba.back.baby.dto;
 
 public record CreateInviteCodeResponse(String inviteCode) {
 }
