+++ back/src/main/java/com/baba/back/baby/exception/RelationGroupNotFoundException.java
@@ -1,4 +1,4 @@
-package com.baba.back.baby.invitation.exception;
+package com.baba.back.baby.exception;
 
 import com.baba.back.exception.NotFoundException;
