+++ back/src/main/java/com/baba/back/baby/repository/InvitationCodeRepository.java
@@ -1,6 +1,6 @@
-package com.baba.back.baby.invitation.repository;
+package com.baba.back.baby.repository;
 
-import com.baba.back.baby.invitation.domain.InvitationCode;
+import com.baba.back.baby.domain.invitation.InvitationCode;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InvitationCodeRepository extends JpaRepository<InvitationCode, Long> {
